[treenode pubkey] follow-up fix for serialize/deserialize pubkeys
