+++ merkle-tree/src/psr_claim.rs
@@ -1,4 +1,5 @@
 use {
+    crate::serde_serialize::pubkey_string_conversion,
     serde::{Deserialize, Serialize},
     solana_program::hash::{Hash, Hasher},
     solana_program::pubkey::Pubkey,
@@ -6,8 +7,11 @@ use {
 
 #[derive(Default, Clone, Eq, Debug, Hash, PartialEq, Deserialize, Serialize)]
 pub struct TreeNode {
+    #[serde(with = "pubkey_string_conversion")]
     pub stake_authority: Pubkey,
+    #[serde(with = "pubkey_string_conversion")]
     pub withdraw_authority: Pubkey,
+    #[serde(with = "pubkey_string_conversion")]
     pub vote_account: Pubkey,
     pub claim: u64,
     pub proof: Option<Vec<[u8; 32]>>,
