+++ snapshot-parser/Cargo.toml
@@ -22,3 +22,4 @@ tokio = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 bincode = { workspace = true }
+merkle-tree = { workspace = true }
